+++ src_ts/components/pages/interventions/intervention-tab-pages/intervention-timing/intervention-timing.ts
@@ -1,8 +1,5 @@
-import {LitElement, customElement, html, property} from 'lit-element';
-import {AnyObject} from '../../../../../types/globals';
-import get from 'lodash-es/get';
+import {LitElement, customElement, html} from 'lit-element';
 import './reporting-requirements/partner-reporting-requirements';
-import cloneDeep from 'lodash-es/cloneDeep';
 import {connect} from 'pwa-helpers/connect-mixin';
 import './intervention-dates/intervention-dates';
 import {getStore} from '../utils/redux-store-access';
@@ -21,15 +18,5 @@ export class InterventionTiming extends connect(getStore())(LitElement) {
       <partner-reporting-requirements class="content-section"> </partner-reporting-requirements>
       <intervention-dates></intervention-dates>
     `;
-  }
-
-  @property({type: Object})
-  intervention!: AnyObject;
-
-  stateChanged(state: any) {
-    // move this to component
-    console.log('--------------------------------------------------------------');
-    const currentIntervention = get(state, 'interventions.current');
-    this.intervention = cloneDeep(currentIntervention);
   }
 }
